+++ src/providers/macroSnapshotContentProvider.ts
@@ -1,0 +1,30 @@
+import * as vscode from 'vscode';
+import { MacroRunInfo } from '../core/execution/macroRunInfo';
+import { ExtensionContext } from '../extensionContext';
+
+export function registerMacroSnapshotContentProvider({
+  runnerManager,
+}: ExtensionContext): vscode.Disposable {
+  return vscode.workspace.registerTextDocumentContentProvider('macro-snapshot', {
+    provideTextDocumentContent(snapshotUri: vscode.Uri) {
+      const uri = vscode.Uri.parse(snapshotUri.path, true);
+      const runner = runnerManager.getRunner(uri);
+      if (runner) {
+        for (const instance of runner.runInstances) {
+          if (instance.id === snapshotUri.fragment) {
+            return instance.snapshot.code;
+          }
+        }
+      }
+      return;
+    },
+  });
+}
+
+export function createSnapshotUri(runInfo: MacroRunInfo) {
+  return vscode.Uri.from({
+    scheme: 'macro-snapshot',
+    path: runInfo.macro.uri.toString(),
+    fragment: runInfo.id,
+  });
+}